+++ Changelog.md
@@ -1,4 +1,5 @@
 Change log:
+* 2022-12-15 - U34.1.1 Rakish Feline, Dashing Feline, Charming Feline
 * 2022-12-08 - U34.1 Festive Yule Goat Kid is now available
 * 2022-11-16 - U34
   * Added 12 pets, some old ones that were forgotten